+++ examples/all_conditions.rs
@@ -7,7 +7,7 @@ use bevy_enhanced_input::prelude::*;
 fn main() {
     // Setup logging to display triggered events.
     let mut log_plugin = LogPlugin::default();
-    log_plugin.filter += ",bevy_enhanced_input::input_context::input_action=trace";
+    log_plugin.filter += ",bevy_enhanced_input::input_action=trace";
 
     App::new()
         .add_plugins((
@@ -23,6 +23,7 @@ struct GamePlugin;
 impl Plugin for GamePlugin {
     fn build(&self, app: &mut App) {
         app.add_input_context::<DummyContext>()
+            .add_observer(binding)
             .add_systems(Startup, spawn);
     }
 }
@@ -31,52 +32,56 @@ fn spawn(mut commands: Commands) {
     commands.spawn(DummyContext);
 }
 
+fn binding(mut trigger: Trigger<Binding<DummyContext>>) {
+    trigger
+        .bind::<PressAction>()
+        .to(PressAction::KEY)
+        .with_conditions(Press::default());
+    trigger
+        .bind::<JustPressAction>()
+        .to(JustPressAction::KEY)
+        .with_conditions(JustPress::default());
+    trigger
+        .bind::<HoldAction>()
+        .to(HoldAction::KEY)
+        .with_conditions(Hold::new(1.0));
+    trigger
+        .bind::<HoldAndReleaseAction>()
+        .to(HoldAndReleaseAction::KEY)
+        .with_conditions(HoldAndRelease::new(1.0));
+    trigger
+        .bind::<PulseAction>()
+        .to(PulseAction::KEY)
+        .with_conditions(Pulse::new(1.0));
+    trigger
+        .bind::<ReleaseAction>()
+        .to(ReleaseAction::KEY)
+        .with_conditions(Release::default());
+    trigger
+        .bind::<TapAction>()
+        .to(TapAction::KEY)
+        .with_conditions(Tap::new(0.5));
+    trigger
+        .bind::<ChordMember1>()
+        .to(ChordMember1::KEY)
+        .with_conditions(BlockBy::<ChordAction>::events_only()); // Don't trigger the action when the chord is active.
+    trigger
+        .bind::<ChordMember2>()
+        .to(ChordMember2::KEY)
+        .with_conditions(BlockBy::<ChordAction>::events_only());
+    trigger.bind::<ChordAction>().with_conditions((
+        Chord::<ChordMember1>::default(),
+        Chord::<ChordMember2>::default(),
+    ));
+    trigger.bind::<BlockerAction>().to(BlockerAction::KEY);
+    trigger
+        .bind::<BlockByAction>()
+        .to(BlockByAction::KEY)
+        .with_conditions(BlockBy::<BlockerAction>::default());
+}
+
 #[derive(Component)]
 struct DummyContext;
-
-impl InputContext for DummyContext {
-    fn context_instance(_world: &World, _entity: Entity) -> ContextInstance {
-        let mut ctx = ContextInstance::default();
-
-        ctx.bind::<PressAction>()
-            .to(PressAction::KEY)
-            .with_conditions(Press::default());
-        ctx.bind::<JustPressAction>()
-            .to(JustPressAction::KEY)
-            .with_conditions(JustPress::default());
-        ctx.bind::<HoldAction>()
-            .to(HoldAction::KEY)
-            .with_conditions(Hold::new(1.0));
-        ctx.bind::<HoldAndReleaseAction>()
-            .to(HoldAndReleaseAction::KEY)
-            .with_conditions(HoldAndRelease::new(1.0));
-        ctx.bind::<PulseAction>()
-            .to(PulseAction::KEY)
-            .with_conditions(Pulse::new(1.0));
-        ctx.bind::<ReleaseAction>()
-            .to(ReleaseAction::KEY)
-            .with_conditions(Release::default());
-        ctx.bind::<TapAction>()
-            .to(TapAction::KEY)
-            .with_conditions(Tap::new(0.5));
-        ctx.bind::<ChordMember1>()
-            .to(ChordMember1::KEY)
-            .with_conditions(BlockBy::<ChordAction>::events_only()); // Don't trigger the action when the chord is active.
-        ctx.bind::<ChordMember2>()
-            .to(ChordMember2::KEY)
-            .with_conditions(BlockBy::<ChordAction>::events_only());
-        ctx.bind::<ChordAction>().with_conditions((
-            Chord::<ChordMember1>::default(),
-            Chord::<ChordMember2>::default(),
-        ));
-        ctx.bind::<BlockerAction>().to(BlockerAction::KEY);
-        ctx.bind::<BlockByAction>()
-            .to(BlockByAction::KEY)
-            .with_conditions(BlockBy::<BlockerAction>::default());
-
-        ctx
-    }
-}
 
 #[derive(Debug, InputAction)]
 #[input_action(output = bool)]
