+++ src/input_condition.rs
@@ -13,8 +13,10 @@ use std::{fmt::Debug, iter};
 
 use bevy::prelude::*;
 
-use super::context_instance::{ActionState, ActionsData};
-use crate::action_value::ActionValue;
+use crate::{
+    action_value::ActionValue,
+    input_context::{ActionState, ActionsData},
+};
 
 pub const DEFAULT_ACTUATION: f32 = 0.5;
 
@@ -25,12 +27,12 @@ pub const DEFAULT_ACTUATION: f32 = 0.5;
 /// or "release" events.
 ///
 /// Can be applied both to inputs and actions.
-/// See [`ActionBind::with_conditions`](super::context_instance::ActionBind::with_conditions)
+/// See [`ActionBind::with_conditions`](super::input_context::ActionBind::with_conditions)
 /// and [`InputBindModCond::with_conditions`](super::input_bind::InputBindModCond::with_conditions).
 pub trait InputCondition: Sync + Send + Debug + 'static {
     /// Returns calculates state.
     ///
-    /// `actions` argument a state of other actions within the currently evaluating context.
+    /// `actions` is a state of other actions within the currently evaluating context.
     fn evaluate(
         &mut self,
         actions: &ActionsData,
@@ -50,7 +52,7 @@ pub trait InputCondition: Sync + Send + Debug + 'static {
 /// on any non-zero value, functioning similarly to a [`Press`](press::Press) condition
 /// with a zero actuation threshold.
 ///
-/// For details about how actions are combined, see [`ContextInstance`](super::context_instance::ContextInstance).
+/// For details about how actions are combined, see [`InputContext`](super::input_context::InputContext).
 pub enum ConditionKind {
     /// The most significant [`ActionState`] from all explicit conditions will be the
     /// resulting state.
@@ -73,7 +75,7 @@ pub enum ConditionKind {
 }
 
 /// Represents collection of bindings that could be passed into
-/// [`ActionBind::with_conditions`](super::context_instance::ActionBind::with_conditions)
+/// [`ActionBind::with_conditions`](super::input_context::ActionBind::with_conditions)
 /// and [`InputBindModCond::with_conditions`](super::input_bind::InputBindModCond::with_conditions).
 pub trait InputConditionSet {
     /// Returns an iterator over conditions.
