+++ src/input_condition/hold_and_release.rs
@@ -3,46 +3,32 @@ use bevy::prelude::*;
 use super::{condition_timer::ConditionTimer, InputCondition, DEFAULT_ACTUATION};
 use crate::{
     action_value::ActionValue,
-    input_context::context_instance::{ActionState, ActionsData},
+    input_context::{ActionState, ActionsData},
 };
 
-/// Returns [`ActionState::Ongoing`] when the input becomes actuated and
-/// [`ActionState::Fired`] when input remained actuated for [`Self::hold_time`] seconds.
+/// Returns [`ActionState::Ongoing`] when input becomes actuated and [`ActionState::Fired`]
+/// when the input is released after having been actuated for [`Self::hold_time`] seconds.
 ///
 /// Returns [`ActionState::None`] when the input stops being actuated earlier than [`Self::hold_time`] seconds.
-/// May optionally fire once, or repeatedly fire.
 #[derive(Clone, Copy, Debug)]
-pub struct Hold {
+pub struct HoldAndRelease {
     /// How long does the input have to be held to cause trigger.
     pub hold_time: f32,
-
-    /// Should this trigger fire only once, or fire every frame once the hold time threshold is met?
-    pub one_shot: bool,
 
     /// Trigger threshold.
     pub actuation: f32,
 
     timer: ConditionTimer,
-
-    fired: bool,
 }
 
-impl Hold {
+impl HoldAndRelease {
     #[must_use]
     pub fn new(hold_time: f32) -> Self {
         Self {
             hold_time,
-            one_shot: false,
             actuation: DEFAULT_ACTUATION,
             timer: Default::default(),
-            fired: false,
         }
-    }
-
-    #[must_use]
-    pub fn one_shot(mut self, one_shot: bool) -> Self {
-        self.one_shot = one_shot;
-        self
     }
 
     #[must_use]
@@ -59,33 +45,29 @@ impl Hold {
     }
 }
 
-impl InputCondition for Hold {
+impl InputCondition for HoldAndRelease {
     fn evaluate(
         &mut self,
         _actions: &ActionsData,
         time: &Time<Virtual>,
         value: ActionValue,
     ) -> ActionState {
-        let actuated = value.is_actuated(self.actuation);
-        if actuated {
-            self.timer.update(time);
+        // Evaluate the updated held duration prior to checking for actuation.
+        // This stops us failing to trigger if the input is released on the
+        // threshold frame due to held duration being 0.
+        self.timer.update(time);
+        let held_duration = self.timer.duration();
+
+        if value.is_actuated(self.actuation) {
+            ActionState::Ongoing
         } else {
             self.timer.reset();
-        }
-
-        let is_first_trigger = !self.fired;
-        self.fired = self.timer.duration() >= self.hold_time;
-
-        if self.fired {
-            if is_first_trigger || !self.one_shot {
+            // Trigger if we've passed the threshold and released.
+            if held_duration >= self.hold_time {
                 ActionState::Fired
             } else {
                 ActionState::None
             }
-        } else if actuated {
-            ActionState::Ongoing
-        } else {
-            ActionState::None
         }
     }
 }
@@ -95,11 +77,11 @@ mod tests {
     use std::time::Duration;
 
     use super::*;
-    use crate::input_context::context_instance::ActionsData;
+    use crate::input_context::ActionsData;
 
     #[test]
-    fn hold() {
-        let mut condition = Hold::new(1.0);
+    fn hold_and_release() {
+        let mut condition = HoldAndRelease::new(1.0);
         let actions = ActionsData::default();
         let mut time = Time::default();
 
@@ -110,16 +92,8 @@ mod tests {
 
         time.advance_by(Duration::from_secs(1));
         assert_eq!(
-            condition.evaluate(&actions, &time, 1.0.into()),
-            ActionState::Fired,
-        );
-        assert_eq!(
-            condition.evaluate(&actions, &time, 1.0.into()),
-            ActionState::Fired,
-        );
-        assert_eq!(
             condition.evaluate(&actions, &time, 0.0.into()),
-            ActionState::None
+            ActionState::Fired
         );
 
         time.advance_by(Duration::ZERO);
@@ -127,21 +101,8 @@ mod tests {
             condition.evaluate(&actions, &time, 1.0.into()),
             ActionState::Ongoing,
         );
-    }
-
-    #[test]
-    fn one_shot() {
-        let mut hold = Hold::new(1.0).one_shot(true);
-        let actions = ActionsData::default();
-        let mut time = Time::default();
-        time.advance_by(Duration::from_secs(1));
-
         assert_eq!(
-            hold.evaluate(&actions, &time, 1.0.into()),
-            ActionState::Fired
-        );
-        assert_eq!(
-            hold.evaluate(&actions, &time, 1.0.into()),
+            condition.evaluate(&actions, &time, 0.0.into()),
             ActionState::None
         );
     }
