+++ src/input_condition/release.rs
@@ -3,7 +3,7 @@ use bevy::prelude::*;
 use super::{InputCondition, DEFAULT_ACTUATION};
 use crate::{
     action_value::ActionValue,
-    input_context::context_instance::{ActionState, ActionsData},
+    input_context::{ActionState, ActionsData},
 };
 
 /// Returns [`ActionState::Ongoing`]` when the input exceeds the actuation threshold and
