+++ src/input_modifier.rs
@@ -11,8 +11,7 @@ use std::{fmt::Debug, iter};
 
 use bevy::prelude::*;
 
-use super::context_instance::ActionsData;
-use crate::action_value::ActionValue;
+use crate::{action_value::ActionValue, input_context::ActionsData};
 
 /// Pre-processor that alter the raw input values.
 ///
@@ -20,7 +19,7 @@ use crate::action_value::ActionValue;
 /// or changing how input maps to axes.
 ///
 /// Can be applied both to inputs and actions.
-/// See [`ActionBind::with_modifiers`](super::context_instance::ActionBind::with_modifiers)
+/// See [`ActionBind::with_modifiers`](super::input_context::ActionBind::with_modifiers)
 /// and [`InputBindModCond::with_modifiers`](super::input_bind::InputBindModCond::with_modifiers).
 pub trait InputModifier: Sync + Send + Debug + 'static {
     /// Returns pre-processed value.
@@ -35,7 +34,7 @@ pub trait InputModifier: Sync + Send + Debug + 'static {
 }
 
 /// Represents collection of bindings that could be passed into
-/// [`ActionBind::with_modifiers`](super::context_instance::ActionBind::with_modifiers)
+/// [`ActionBind::with_modifiers`](super::input_context::ActionBind::with_modifiers)
 /// and [`InputBindModCond::with_modifiers`](super::input_bind::InputBindModCond::with_modifiers).
 pub trait InputModifierSet {
     /// Returns an iterator over modifiers.
