+++ src/input_modifier/dead_zone.rs
@@ -1,7 +1,7 @@
 use bevy::prelude::*;
 
 use super::InputModifier;
-use crate::{action_value::ActionValue, input_context::context_instance::ActionsData};
+use crate::{action_value::ActionValue, input_context::ActionsData};
 
 /// Remaps input values within the range [Self::lower_threshold] to [Self::upper_threshold] onto the range 0 to 1.
 /// Values outside this range are clamped.
